Do not sleep before retry on deadline exceeded
Previously, RetryingCallable exponentially backs off if it gets
an error.
If the error is DEADLINE_EXCEEDED, it should continue immediately,
since we have already been waiting for long enough.

This commit makes RetryingCallable wait for 1ms after getting
DEADLINE_EXCEEDED error. The 1ms wait time is introduced so that
it plays nice with our fake clock used for testing.