WIP: bundling: optimize partion key
Partition key is used in bundling methods to determine
whether two method calls can be merged into one RPC request.
Previously this key is a string.
This is inefficient, as a new string needs to be created
every time a bundling method is called.

This commit replaces the string with the PartitionKey type.
Since protobuf objects have value-equality,
PartitionKey instances simply compare equal if their
constituents are equal and in the same order.

In a simple load-test, this decreases the time required to
write 100,000 log messages to Stackdriver from
9.6s to 5.4s and reduces CPU consumption from 136% to 117%.

I have not compared this with Pubsub's implementation,
so it is possible that we might need to change bundling
at a more fundamental level.
Posting this PR for early feedback and to safeguard against
my machine dying.