Update the ResumptionStrategy interface to allow it to modify the incoming response stream.
The general use cases for this are:

1. If a stream can only be resumed from a particular point, then its strategy will resume from the closest point, suppressing the repeated messages in the retry
2. If the stream elements have sequence numbers relative to the start of the stream, then strategy can re-write the sequence numbers when it performs a retry.

The specific use case is to enable retries for Bigtable's MutateRows RPC. Which has a batch request and streamed responses. Each entry can fail independently of the others. I would like to reuse the streaming retry infrastructure to implement retries. On failure, the strategy will find all of the failed subresponses and compose a new request with them. The responses will have to be rewritten to adjust indexes on the response entries.