fix(batcher): fix logged messge when exception occurs post processing
This commit addresses logged message while batch post
processing.

Currently, if a batch is successful and some exception
occurs during processing then the Batcher generates an exception message
with "Batching finished with 3 batches failed" which is not an accurate
message.

After this commit, for the same scenario, the message will be
logged with "Batching finished with 3 partial failures...".