Adds ability to transform exceptions returned by ApiExceptionFactory.
Defines a ServiceProvider interface ExceptionTransformation which can be
used to define a custom implementation of a transformation that should
be applied to exceptions before returning them from the factory.

This is used by the Ads API Java client library to extract a binary proto
message from the gRPC trailers. Currently we post-process the generated code
to override the callable factory. Each callable returned has to be hard coded
to generate a GoogleAdsException instead of generic exceptions. The approach
in this change is preferable since (1) we don't need to override N different
callable types and (2) it can be reused across different client library
implementations.